Rework extension lifecycle (#21)
* Remove the cookie extension

This extension currently does not expose any API to javascript. Its
only purpose is to inject cookies at the creation of the webview.
Instead of complicating the extension lifecycle to support this we
will investigate alternative methods for allowing an extension to
participate in the initialization of the webview.

* Remove asynchronous extension initialization

Simplified extension initialization to the bare minimum for now.
Initialization is now synchronous and is primarily there to support
binding native methods to the webview. Alternative initialization
hooks will be added later to restore mechanisms for injecting cookies
or doing other pre-webview creation setup.

* Remove some dead code